Final big clean on script before cleaning up the comments and submitting.
I fixed the whole import section so that the data is downloaded to the user's machine in a consistent manner from a URL and all of the file writing is done as efficiently as possible.

In the new line 30 I found an easier way to add names from a vector.  I thought I had to concacenate the vector, but instead I just pass it directly to the colnames function.

I couldn't figure out how to use janitor:clean_names() on the dataset in a way that allowed me to easily separate the variable measurements from the averages of different factors.

In line 49 I've added the necessary code to meet requirement 3 which is "uses descriptive names to name the activities in the data set"

In lines 53 to 56 I address the final prompt 5 "From the data set in step 4, creates a second, independent tidy data set with the average of each variable for each activity and each subject." and finished the summary table by grouping by different variables in appropriate order and then tidy-ing by spreading the "mean" and "std" in "Stat_Type" into separate columns with the newly calculated "Avg_val" as their observations

Finally, I wrote the Summary Table into the folder "outputs" as "summary_dataset.txt"